// src/api/generated/models/appUserResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Gaula Digital — API
 * Sistema de información operacional del GAULA, Ejército Nacional de Colombia.
 * OpenAPI spec version: v1
 */
import type { AppUserResponseClearanceLevel } from './appUserResponseClearanceLevel';

export interface AppUserResponse {
  id?: string;
  militaryId?: string;
  rank?: string;
  firstName?: string;
  lastName?: string;
  operationalUnitId?: string;
  territorialUnitId?: string;
  clearanceLevel?: AppUserResponseClearanceLevel;
  active?: boolean;
  lastLoginAt?: string;
  roles?: string[];
}