// src/api/generated/models/pageResponseAppUserResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Gaula Digital — API
 * Sistema de información operacional del GAULA, Ejército Nacional de Colombia.
 * OpenAPI spec version: v1
 */
import type { AppUserResponse } from './appUserResponse';

export interface PageResponseAppUserResponse {
  content?: AppUserResponse[];
  totalElements?: number;
  totalPages?: number;
  pageNumber?: number;
  pageSize?: number;
}